# backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/favorites/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "favorites-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/products/stats/summary": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/attributes/stats/summary": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-attributes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/categories/stats/summary": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/notifications/stats/overview": {
      "get": {
        "operationId": "getStats",
        "summary": "Get notification statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          }
        },
        "tags": [
          "notifications-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/support/tickets": {
      "post": {
        "operationId": "createTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "getMyTickets",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/support/tickets/{id}": {
      "get": {
        "operationId": "getTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/support/tickets/{id}/messages": {
      "get": {
        "operationId": "getTicketMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSupportMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/support/tickets": {
      "get": {
        "operationId": "getAllTickets",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "in_progress",
                "waiting_for_user",
                "resolved",
                "closed"
              ],
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "low",
                "medium",
                "high",
                "urgent"
              ],
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "technical",
                "billing",
                "products",
                "services",
                "account",
                "other"
              ],
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/support/tickets/{id}": {
      "get": {
        "operationId": "getTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "updateTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/support/tickets/{id}/messages": {
      "get": {
        "operationId": "getTicketMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Ticket ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSupportMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/support/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "support-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/services/requests/{id}/offers": {
      "get": {
        "operationId": "getOffers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "services-customer"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/services/requests/{id}": {
      "get": {
        "operationId": "getRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "services-admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/media/stats/summary": {
      "get": {
        "operationId": "getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tagadod API",
    "description": "Complete e-commerce platform for solar products with advanced analytics and multi-language support",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.tagadod.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateSupportTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "عنوان المشكلة",
            "example": "مشكلة في لوحة الطاقة الشمسية"
          },
          "description": {
            "type": "string",
            "description": "وصف تفصيلي للمشكلة",
            "example": "اللوحة لا تعمل بشكل صحيح منذ يومين"
          },
          "category": {
            "type": "string",
            "description": "تصنيف المشكلة",
            "enum": [
              "technical",
              "billing",
              "products",
              "services",
              "account",
              "other"
            ],
            "example": "technical"
          },
          "priority": {
            "type": "string",
            "description": "أولوية المشكلة",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ],
            "example": "medium"
          },
          "attachments": {
            "description": "روابط الملفات المرفوعة",
            "example": [
              "https://cdn.bunny.net/uploads/image1.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "description": "بيانات إضافية (مثل رقم الطلب أو رقم المنتج)",
            "example": {
              "orderId": "12345",
              "productId": "67890"
            }
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "AddSupportMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "محتوى الرسالة",
            "example": "شكراً لك على الإبلاغ عن المشكلة. سنعمل على حلها في أقرب وقت ممكن."
          },
          "attachments": {
            "description": "روابط الملفات المرفوعة مع الرسالة",
            "example": [
              "https://cdn.bunny.net/uploads/solution.pdf"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isInternal": {
            "type": "boolean",
            "description": "رسالة داخلية (للمشرفين فقط)",
            "example": false
          },
          "metadata": {
            "type": "object",
            "description": "بيانات إضافية",
            "example": {
              "action": "escalated"
            }
          }
        },
        "required": [
          "content"
        ]
      },
      "UpdateSupportTicketDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "حالة التذكرة",
            "enum": [
              "open",
              "in_progress",
              "waiting_for_user",
              "resolved",
              "closed"
            ],
            "example": "in_progress"
          },
          "priority": {
            "type": "string",
            "description": "أولوية التذكرة",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ],
            "example": "high"
          },
          "category": {
            "type": "string",
            "description": "تصنيف التذكرة",
            "enum": [
              "technical",
              "billing",
              "products",
              "services",
              "account",
              "other"
            ],
            "example": "technical"
          },
          "assignedTo": {
            "type": "string",
            "description": "معرف المشرف المسؤول",
            "example": "507f1f77bcf86cd799439011"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "description": "تاريخ الحل (يتم تعيينه تلقائياً عند تغيير الحالة إلى RESOLVED)"
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "description": "تاريخ الإغلاق (يتم تعيينه تلقائياً عند تغيير الحالة إلى CLOSED)"
          }
        }
      },
...
    }
  }
}